% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_fix_path.R
\name{addin_fix_path}
\alias{addin_fix_path}
\title{RStudio addin: fix Windows paths}
\usage{
addin_fix_path()
}
\description{
Replaces backslashes with forward slashes in Windows-style paths
(\verb{C:\\...} or \verb{\\\\server\\...}). Operates on the selected text, or the
entire active file if nothing is selected.
}
\keyword{internal}
