% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_fix_path.R
\name{addin_flip_backslash}
\alias{addin_flip_backslash}
\title{RStudio addin: flip backslashes in selection}
\usage{
addin_flip_backslash()
}
\description{
Replaces all backslashes with forward slashes in the selected text.
Requires a selection.
}
\keyword{internal}
