% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ta.R
\name{read_ta}
\alias{read_ta}
\title{Read a Greenlandic TA file}
\usage{
read_ta(ta)
}
\arguments{
\item{ta}{Path to the TA file.}
}
\value{
A tibble with the columns described above.
}
\description{
Reads a Greenlandic TA file. Supports current prices (L), constant
prices (F), and prior-year prices (D) file types.
}
\details{
Column positions (0-based byte offsets):\tabular{lll}{
   Column \tab Start \tab End \cr
   nrnr \tab 0 \tab 5 \cr
   trans \tab 6 \tab 12 \cr
   brch \tab 13 \tab 17 \cr
   bas \tab 18 \tab 31 \cr
   eng \tab 32 \tab 43 \cr
   det \tab 44 \tab 55 \cr
   afg \tab 56 \tab 67 \cr
   moms \tab 68 \tab 79 \cr
   kbx \tab 80 \tab 91 \cr
   prim \tab 92 \tab 103 \cr
   afstm \tab 104 \tab 107 \cr
   fval \tab 108 \tab end \cr
}


Columns \code{nrnr}, \code{trans}, \code{brch}, \code{afstm}, and \code{fval} are read as
character; all others as double.
}
\examples{
\dontrun{
df <- read_ta("ta.file")
}

}
