% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ta.R
\name{read_ta}
\alias{read_ta}
\title{Read a Greenlandic TA file}
\usage{
read_ta(ta)
}
\arguments{
\item{ta}{Path to the TA file.}
}
\value{
A tibble with the columns described above.
}
\description{
Reads a Greenlandic TA file. Supports current prices (L), constant
prices (F), and prior-year prices (D) file types.
}
\details{
Column positions (0-based byte offsets):\tabular{lll}{
   Column \tab Start \tab End \cr
   nrnr \tab 0 \tab 5 \cr
   trans \tab 6 \tab 12 \cr
   brch \tab 13 \tab 17 \cr
   bas \tab 18 \tab 32 \cr
   eng \tab 33 \tab 45 \cr
   det \tab 46 \tab 58 \cr
   afg \tab 59 \tab 71 \cr
   moms \tab 72 \tab 84 \cr
   kbx \tab 85 \tab 97 \cr
   prim \tab 98 \tab 103 \cr
   afstm \tab 104 \tab 107 \cr
   fval \tab 108 \tab end \cr
}


Columns \code{nrnr}, \code{trans}, \code{brch}, \code{afstm}, and \code{fval} are read as
character; all others as double.
}
\examples{
\dontrun{
df <- read_ta("ta.file")
}

}
